fix get all rates

+++ turradgiver-bal/Dtos/Ads/SearchDto.cs
@@ -21,7 +21,7 @@ namespace turradgiver_bal.Dtos.Ads
             }
         }
 
-        [Range(1, float.MaxValue, ErrorMessage = "Please enter valid page Number")]
+        [Range(1, int.MaxValue, ErrorMessage = "Please enter valid page Number")]
         public int Page { get; set; } = 1;
     }
 }
